course list and form working! :godmode:
It turns out that `type` is a reserved word (which should've been
obvious but whatever), so I changed type to course_type, besides
that the rest is pretty straight forward. New template, new
controller and actions. The front end side is starting to become
a monster, when I have something working I should pause and
restructure the whole app.